refactor: Move areas to organisms
