+++ src/components/Organisms/JSONTranslationArea.js
@@ -1,5 +1,5 @@
 import {Container} from "@material-ui/core";
-import JSONCreator from "./JSONCreator";
+import JSONCreator from "../Molecules/JSONCreator";
 import React from "react";
 
 export const JSONTranslationArea = () => {
